Version 0.0.5
Some updates to the JS and UX.
Dev is being halted until the client gets back - this could be a while.
Opportunity to do some code refactoring; the layout's good but it could get unwieldy if I continue in the current direction. As Danny mentioned, maybe making this more knockout heavy and state based will optimize the interactive. TBD.